requirements.txt
streamlit
python-dotenv
bcrypt
requests
openai
arcadepy
sqlalchemy
psycopg2-binary
langchain
langchain-community
langchain-openai
pandas
plotly
prompt_toolkit>=3.0.0
rich>=13.0.0
playwright
python-fasthtml
langgraph
python-pptx
pymupdf4llm>=0.0.17
python-docx>=1.1.0
fastembed>=0.5.0
langchain-text-splitters>=0.3.0
